utils: Fix SC2115 and ignore SC2034 for array

+++ utils/clean.sh
@@ -5,5 +5,6 @@ REPOSITORY="$(pwd)"
 source "$(pwd)/utils/arrays.sh"
 
 for i in "${!files[@]}"; do
-    rm -rf "${REPOSITORY}/${i}"
+    rm -rf "${REPOSITORY:?}/${i}"
+
 done
